tweak

+++ opt.py
@@ -640,7 +640,7 @@ class OptWBoundEignVal(object):
                 if 'max' in self.test_func:
                     _, predicted = torch.max(ops.data, 1)
                 else:
-                    predicted = (ops.data > 0.5).int()
+                    predicted = (ops.data > 0.5).float()
                 target = target.to(self.device)
                 if 'acc' in self.test_func:
                     print(predicted)
